@@ -2,19 +2,20 @@
   "name": "@cycjimmy/mini-xhr",
   "version": "4.1.0",
   "description": "A very lightweight javascript library for HTTP requests.",
+  "type": "module",
   "main": "dist/mini-xhr.cjs.js",
   "module": "dist/mini-xhr.esm.js",
   "browser": "dist/mini-xhr.umd.min.js",
   "scripts": {
     "start": "npm run dev",
-    "test": "jest --color --forceExit --detectOpenHandles",
-    "coveralls": "jest --color --forceExit --detectOpenHandles --coverage",
-    "lint": "eslint -c .eslintrc 'src/**/*.js'",
-    "lint:fix": "eslint --fix -c .eslintrc 'src/**/*.js' --fix",
-    "dev": "cross-env NODE_ENV=development rollup -c rollup/rollup.config.dev.js --watch",
+    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest --color --forceExit --detectOpenHandles",
+    "coveralls": "cross-env NODE_OPTIONS=--experimental-vm-modules jest --color --forceExit --detectOpenHandles --coverage",
+    "lint": "eslint --ext .js,.cjs,.mjs -c .eslintrc .",
+    "lint:fix": "eslint --ext .js,.cjs,.mjs --fix -c .eslintrc . --fix",
+    "dev": "cross-env NODE_ENV=development rollup -c rollup/rollup.config.dev.mjs --watch",
     "build": "npm run build:prod:umd && npm run build:prod",
-    "build:prod": "cross-env NODE_ENV=production rollup -c rollup/rollup.config.prod.js",
-    "build:prod:umd": "cross-env NODE_ENV=production rollup -c rollup/rollup.config.dev.js"
+    "build:prod": "cross-env NODE_ENV=production rollup -c rollup/rollup.config.prod.mjs",
+    "build:prod:umd": "cross-env NODE_ENV=production rollup -c rollup/rollup.config.dev.mjs"
   },
   "repository": {
     "type": "git",
@@ -38,24 +39,26 @@
     "access": "public"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.3",
+    "@babel/core": "^7.17.7",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.12.1",
-    "@cycjimmy/awesome-js-funcs": "^2.4.0",
-    "@cycjimmy/config-lib": "^1.7.1",
-    "@rollup/plugin-babel": "^5.2.1",
-    "@rollup/plugin-commonjs": "^16.0.0",
+    "@babel/preset-env": "^7.16.11",
+    "@cycjimmy/awesome-js-funcs": "^4.0.4",
+    "@cycjimmy/config-lib": "^3.0.5",
+    "@jest/globals": "^27.5.1",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-commonjs": "^21.0.2",
+    "@rollup/plugin-eslint": "^8.0.1",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^10.0.0",
-    "babel-jest": "^26.6.3",
-    "coveralls": "^3.1.0",
-    "cross-env": "^7.0.2",
-    "eslint": "^7.12.1",
-    "eslint-config-rollup": "^0.1.0",
-    "jest": "^26.6.3",
-    "rollup": "^2.33.1",
-    "rollup-plugin-browsersync": "^1.3.1",
-    "rollup-plugin-eslint": "^7.0.0",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "babel-jest": "^27.5.1",
+    "coveralls": "^3.1.1",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.11.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.25.4",
+    "jest": "^27.5.1",
+    "rollup": "^2.70.1",
+    "rollup-plugin-browsersync": "^1.3.3",
     "rollup-plugin-terser": "^7.0.2"
   }
 }
