@@ -1,10 +1,11 @@
-module.exports = {
+export default {
+  testEnvironment: 'jsdom',
   coverageThreshold: {
     global: {
       branches: 50,
       functions: 50,
       lines: 50,
       statements: 50,
-    }
+    },
   },
 };
