@@ -1,5 +1,5 @@
 export const responseSuccess = JSON.stringify({
-  data: 'success'
+  data: 'success',
 });
 
 export const testUrl = 'https://aa.bb.test.com/cycjimmy';
@@ -7,4 +7,3 @@ export const testUrl = 'https://aa.bb.test.com/cycjimmy';
 export const statusTextFor404 = 'Not Found';
 
 export const JsonpErrorTextForTimeout = 'timeout';
-
