@@ -1,11 +1,12 @@
+/* eslint no-undef: off */
 import createXhrMock from '../__mocks__/createXhrMock';
-import {responseSuccess, testUrl, statusTextFor404} from '../__mocks__/someTestVariables';
+import { responseSuccess, testUrl, statusTextFor404 } from '../__mocks__/someTestVariables';
 import _XMLHttpRequest from '../src/_XMLHttpRequest';
 
 describe('_XMLHttpRequest', () => {
   test('_XMLHttpRequest default', (done) => {
     createXhrMock({
-      response: responseSuccess
+      response: responseSuccess,
     });
 
     _XMLHttpRequest({
@@ -13,15 +14,15 @@ describe('_XMLHttpRequest', () => {
       success: (data) => {
         expect(data).toBe(responseSuccess);
         done();
-      }
+      },
     });
   });
 
-  test('_XMLHttpRequest timeout', done => {
+  test('_XMLHttpRequest timeout', (done) => {
     createXhrMock({
       readyState: 4,
       status: 0,
-      response: responseSuccess
+      response: responseSuccess,
     });
 
     _XMLHttpRequest({
@@ -32,34 +33,34 @@ describe('_XMLHttpRequest', () => {
     });
   });
 
-  test('_XMLHttpRequest fail', done => {
+  test('_XMLHttpRequest fail', (done) => {
     createXhrMock({
       readyState: 4,
       status: 404,
-      statusText: statusTextFor404
+      statusText: statusTextFor404,
     });
 
     _XMLHttpRequest({
       url: testUrl,
-      fail: (err) => {
+      fail: () => {
         done();
-      }
+      },
     });
   });
 
   test('_XMLHttpRequest cover success function', (done) => {
     createXhrMock({
-      response: responseSuccess
+      response: responseSuccess,
     });
 
     _XMLHttpRequest({
       url: testUrl,
       data: {
-        a: 1
+        a: 1,
       },
       headers: {
-        h1: 1
-      }
+        h1: 1,
+      },
     });
 
     setTimeout(done, 1e3);
@@ -69,7 +70,7 @@ describe('_XMLHttpRequest', () => {
     createXhrMock({
       readyState: 4,
       status: 404,
-      statusText: statusTextFor404
+      statusText: statusTextFor404,
     });
 
     _XMLHttpRequest({
