@@ -1,15 +1,19 @@
+/* eslint no-undef: off */
+/* eslint import/no-named-as-default-member: off */
 import createXhrMock from '../__mocks__/createXhrMock';
-import {responseSuccess, testUrl, statusTextFor404, JsonpErrorTextForTimeout} from '../__mocks__/someTestVariables';
+import {
+  responseSuccess, testUrl, statusTextFor404, JsonpErrorTextForTimeout,
+} from '../__mocks__/someTestVariables';
 import miniXhr from '../src/index';
 
 describe('miniXhr', () => {
   test('miniXhr.get() default', () => {
     createXhrMock({
-      response: responseSuccess
+      response: responseSuccess,
     });
 
     return miniXhr.get(testUrl)
-      .then(data => {
+      .then((data) => {
         expect(data).toBe(responseSuccess);
       });
   });
@@ -18,11 +22,11 @@ describe('miniXhr', () => {
     createXhrMock({
       readyState: 4,
       status: 404,
-      statusText: statusTextFor404
+      statusText: statusTextFor404,
     });
 
     return miniXhr.get(testUrl)
-      .catch(err => {
+      .catch((err) => {
         expect(err.status).toBe(404);
         expect(err.statusText).toBe(statusTextFor404);
       });
@@ -30,11 +34,11 @@ describe('miniXhr', () => {
 
   test('miniXhr.post() default', () => {
     createXhrMock({
-      response: responseSuccess
+      response: responseSuccess,
     });
 
     return miniXhr.post(testUrl)
-      .then(data => {
+      .then((data) => {
         expect(data).toBe(responseSuccess);
       });
   });
@@ -43,11 +47,11 @@ describe('miniXhr', () => {
     createXhrMock({
       readyState: 4,
       status: 404,
-      statusText: statusTextFor404
+      statusText: statusTextFor404,
     });
 
     return miniXhr.post(testUrl)
-      .catch(err => {
+      .catch((err) => {
         expect(err.status).toBe(404);
         expect(err.statusText).toBe(statusTextFor404);
       });
@@ -58,13 +62,11 @@ describe('miniXhr', () => {
     setTimeout(done, 1e3);
   });
 
-  test('miniXhr.jsonp() works when timeout', () => {
-    return miniXhr.jsonp(testUrl, {
-      data: {a: 1},
-      timeout: 2e3,
-    })
-      .catch(err => {
-        expect(err).toBe(JsonpErrorTextForTimeout);
-      });
-  });
+  test('miniXhr.jsonp() works when timeout', () => miniXhr.jsonp(testUrl, {
+    data: { a: 1 },
+    timeout: 2e3,
+  })
+    .catch((err) => {
+      expect(err).toBe(JsonpErrorTextForTimeout);
+    }));
 });
