@@ -1,4 +1,5 @@
-import {dataStringMakeUp, getGlobal} from '../src/tools';
+/* eslint no-undef: off */
+import { dataStringMakeUp, getGlobal } from '../src/tools';
 
 describe('tools', () => {
   test('dataStringMakeUp() default test', () => {
@@ -6,11 +7,11 @@ describe('tools', () => {
   });
 
   test('dataStringMakeUp({a: 1, b: 2}) return a=1&b=2', () => {
-    expect(dataStringMakeUp({a: 1, b: 2})).toBe('a=1&b=2');
+    expect(dataStringMakeUp({ a: 1, b: 2 })).toBe('a=1&b=2');
   });
 
   test('dataStringMakeUp() cover test', () => {
-    expect(dataStringMakeUp(Object.create({a: 1, b: 2}))).toBe('');
+    expect(dataStringMakeUp(Object.create({ a: 1, b: 2 }))).toBe('');
   });
 
   test('getGlobal() test return window', () => {
