@@ -1,19 +1,20 @@
-import {eslint} from 'rollup-plugin-eslint';
+/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
+/* eslint import/extensions: ["error", "ignorePackages", {"js": off}] */
+import eslint from '@rollup/plugin-eslint';
 import json from '@rollup/plugin-json';
 import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
-import babel from '@rollup/plugin-babel';
-import {terser} from 'rollup-plugin-terser';
-import pkg from '../package.json';
-import myBanner from '@cycjimmy/config-lib/chore/myBanner';
-import midlineToCamel from '@cycjimmy/awesome-js-funcs/string/midlineToCamel';
+import { babel } from '@rollup/plugin-babel';
+import { terser } from 'rollup-plugin-terser';
 
-// config
-import terserOption from '@cycjimmy/config-lib/terser/4.x/production';
+import myBanner from '@cycjimmy/config-lib/esm/chore/myBanner.js';
+import midlineToCamel from '@cycjimmy/awesome-js-funcs/esm/string/midlineToCamel.js';
+import terserOption from '@cycjimmy/config-lib/esm/terser/4.x/production.js';
+
+import pkg from './package.cjs';
 
 export const IS_DEVELOPMENT = process.env.NODE_ENV === 'development';
 export const IS_PRODUCTION = process.env.NODE_ENV === 'production';
-export const IS_DEPLOYMENT = process.env.NODE_ENV === 'deployment';
 
 export const input = './src/index.js';
 export const name = midlineToCamel(pkg.name.replace(/.+\//g, ''));
@@ -26,7 +27,7 @@ export const plugins = [
     fix: true,
     exclude: [
       '**/*.(css|scss)',
-    ]
+    ],
   }),
   resolve(),
   babel({ babelHelpers: 'bundled' }),
@@ -34,4 +35,3 @@ export const plugins = [
 ];
 
 export const terserPlugins = (IS_PRODUCTION && terser(terserOption));
-
